RSG_Integration_2/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
lxml==4.9.3
zeep==4.2.1
openpyxl==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1